event_management_system_final/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow>=10.0.0
qrcode[pil]>=7.4.0
openpyxl>=3.1.0